requirements.txt
pytest-playwright
pytest-html
pytest-metadata